requirements.txt
pandas==2.2.1
google-cloud-bigquery==3.18.0
google-generativeai==0.4.1
Faker==24.4.0
python-dotenv==1.0.1
